build: moves logic to on wallet connect adds created_at to intercom update
build: addresses comment
