+++ package.json
@@ -17,6 +17,7 @@
     "@ethersproject/abi": "^5.7.0",
     "@ethersproject/bytes": "^5.7.0",
     "@ethersproject/providers": "^5.7.1",
+    "@intercom/messenger-js-sdk": "^0.0.14",
     "@lottiefiles/react-lottie-player": "3.5.3",
     "@mui/base": "^5.0.0-beta.40",
     "@mui/icons-material": "^5.16.6",
@@ -82,7 +83,6 @@
     "react-scroll": "^1.9.0",
     "react-slick": "^0.29.0",
     "react-swipeable": "^7.0.0",
-    "react-use-intercom": "^5.4.0",
     "react-youtube": "^9.0.2",
     "recharts": "^2.3.2",
     "regenerator-runtime": "0.13.7",
