+++ packages/shared/wallet/src/hooks/index.ts
@@ -1,3 +1,2 @@
 export * from './use-wallet-side-drawer';
-export * from './use-intercom-update';
 export * from './use-connect';
