+++ packages/shared/wallet/src/wallet-selector/wallet-selector.tsx
@@ -11,7 +11,7 @@ import {
   CopyCaption,
 } from '@notional-finance/mui';
 import { useSideDrawerManager } from '@notional-finance/notionable-hooks';
-import { useWalletSideDrawer, useConnect, useIntercomUpdate } from '../hooks';
+import { useWalletSideDrawer, useConnect } from '../hooks';
 import {
   PORTFOLIO_ACTIONS,
   PORTFOLIO_CATEGORIES,
@@ -40,7 +40,6 @@ export function WalletSelector() {
   const [showAlert, setShowAlert] = useState<boolean>(false);
   const { setWalletSideDrawer, clearWalletSideDrawer } = useSideDrawerManager();
   const { openDrawer } = useWalletSideDrawer();
-  useIntercomUpdate();
 
   useEffect(() => {
     getNotificationsData();
