+++ packages/util/src/enums.ts
@@ -165,6 +165,7 @@ export enum TRACKING_EVENTS {
   REJECT_TXN = 'RejectTxn',
   CONFIRMATION = 'Confirmation',
   DISCONNECT_WALLET = 'DisconnectWallet',
+  WALLET_CONNECTED = 'WalletConnected',
 }
 
 export type SIDE_DRAWERS_TYPE = SETTINGS_SIDE_DRAWERS | PORTFOLIO_ACTIONS;
